% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PackageDescription.R
\docType{package}
\name{cvmdisc}
\alias{cvmdisc}
\alias{cvmdisc-package}
\title{Fit Grouped Data}
\description{
cvmdisc is a package for fitting data to binomial, poisson, and grouped continuous distributions as well as testing their Goodness of Fit using Cramer-von-Mises and Anderson-Darling statistics
}
\details{
The cvmdisc package includes a function that fits and tests grouped/discrete data to multiple distributions, as well as multiple auxillary functions called by the main function.
}
\section{Distributions}{


The Binomial, Poisson, Exponential, Gamma, Log-normal, Normal, Uniform and Weibull distributions are currently supported in the cvmdisc package.
}

\seealso{
\code{\link{groupFit}}: Fit data and test GoF

\code{\link{distrFit}}: Fit data

\code{\link{cvmTest}}: Compute test statistics

\code{\link{cvmPval}}: Compute p values
}
